split up test files for state machine class
